+++ onnxruntime/core/graph/model.cc
@@ -387,14 +387,14 @@ Status Model::Save(Model& model, const std::wstring& file_path) {
 }
 #endif
 
-Status Model::Load(const std::basic_string<ORTCHAR_T>& file_path,
+Status Model::Load(const PathString& file_path,
                    ONNX_NAMESPACE::ModelProto& model_proto) {
   return LoadModel(file_path, model_proto);
 }
 
 GSL_SUPPRESS(r .30)  // spurious warnings. p_model is potentially reset in the internal call to Load
 GSL_SUPPRESS(r .35)
-Status Model::Load(const std::basic_string<ORTCHAR_T>& file_path, std::shared_ptr<Model>& p_model,
+Status Model::Load(const PathString& file_path, std::shared_ptr<Model>& p_model,
                    const IOnnxRuntimeOpSchemaRegistryList* local_registries,
                    const logging::Logger& logger) {
   return LoadModel(file_path, p_model, local_registries, logger);
