+++ onnxruntime/core/graph/model.h
@@ -126,11 +126,11 @@ class Model {
 
   static common::Status Load(std::istream& model_istream, ONNX_NAMESPACE::ModelProto* p_model_proto);
 
-  static common::Status Load(const std::basic_string<ORTCHAR_T>& file_path,
+  static common::Status Load(const PathString& file_path,
                              /*out*/ ONNX_NAMESPACE::ModelProto& model_proto);
 
   // TODO(Task:132) Use of shared_ptr<X>* in Load/Save methods is confusing.
-  static common::Status Load(const std::basic_string<ORTCHAR_T>& file_path,
+  static common::Status Load(const PathString& file_path,
                              /*out*/ std::shared_ptr<Model>& p_model,
                              const IOnnxRuntimeOpSchemaRegistryList* local_registries,
                              const logging::Logger& logger);
