+++ onnxruntime/test/optimizer/initializer_test.cc
@@ -14,43 +14,18 @@
 #include "core/common/common.h"
 #include "core/framework/endian_utils.h"
 #include "test/util/include/asserts.h"
+#include "test/util/include/file_util.h"
 
 namespace onnxruntime {
 namespace test {
 namespace {
-class FileDeleter {
- public:
-  FileDeleter() = default;
-  FileDeleter(const PathString& path) : path_{path} {}
-  FileDeleter(FileDeleter&& other) { *this = std::move(other); }
-  FileDeleter& operator=(FileDeleter&& other) {
-    CleanUp();
-    path_ = std::move(other.path_);
-    other.path_.clear();
-    return *this;
-  }
-  ~FileDeleter() { CleanUp(); }
-
- private:
-  ORT_DISALLOW_COPY_AND_ASSIGNMENT(FileDeleter);
-
-  void CleanUp() {
-    if (!path_.empty()) {
-      std::remove(ToMBString(path_).c_str());
-      path_.clear();
-    }
-  }
-
-  PathString path_;
-};
-
 template <typename T>
-Status WriteExternalDataFile(gsl::span<const T> data, const PathString& path, FileDeleter& file_deleter) {
+Status WriteExternalDataFile(gsl::span<const T> data, const PathString& path, ScopedFileDeleter& file_deleter) {
   std::vector<char> data_bytes(data.size_bytes());
   ORT_RETURN_IF_ERROR(onnxruntime::utils::WriteLittleEndian(data, gsl::make_span(data_bytes)));
   std::ofstream out{path, std::ios::binary | std::ios::trunc};
   ORT_RETURN_IF_NOT(out && out.write(data_bytes.data(), data_bytes.size()));
-  file_deleter = FileDeleter{path};
+  file_deleter = ScopedFileDeleter{path};
   return Status::OK();
 }
 
@@ -76,7 +51,7 @@ TEST(OptimizerInitializerTest, LoadExternalData) {
   const gsl::span<const int> tensor_data_span = gsl::make_span(tensor_data);
   const auto tensor_data_dir_path = Path::Parse(ToPathString("."));
   const auto tensor_data_dir_relative_path = Path::Parse(ToPathString("OptimizerInitializerTest_LoadExternalData.bin"));
-  FileDeleter file_deleter{};
+  ScopedFileDeleter file_deleter{};
 
   ASSERT_STATUS_OK(WriteExternalDataFile(
       tensor_data_span, (tensor_data_dir_path / tensor_data_dir_relative_path).ToPathString(), file_deleter));
